Change message about talk length on request page
The instructions used to tell people to select a label for "Talk idea"
or "Lightning Talk Idea" but not all users can add a label to the
issue. Plus "Talk Idea"s can be of varying length and we want to know
how long a "Talk Idea" will be when we want to schedule a meeting.

Address both issues by asking the user to include a talk length when
submitting an issue this means they don't have to add a label and we
(hopefully) get fuller information about time.